Rename views' files to match naming convention
