+++ src/routes/Routes.js
@@ -12,19 +12,19 @@ import fusTheme from '../components/fusTheme'
 import SideNav from '../components/SideNav'
 
 /* Import Views */
-import Glossary from '../views/Glossary'
+import Glossary from '../views/glossary'
 import Home from '../views/Home'
-import Letterhead from '../views/Letterhead'
+import Letterhead from '../views/letterhead'
 import Logos from '../views/Logos'
-import LogosPosters from '../views/LogosPosters'
-import NotFound from '../views/NotFound'
+import LogosPosters from '../views/logosPosters'
+import NotFound from '../views/notFound'
 import PlanningGuide from '../views/PlanningGuide'
 import Posters from '../views/Posters'
-import PosterVideos from '../views/PosterVideos'
+import PosterVideos from '../views/posterVideos'
 import ScrollIntoView from '../components/ScrollIntoView'
 import ServiceRequest from '../views/ServiceRequest'
-import Services from '../views/Services'
-import Story from '../views/Story'
+import Services from '../views/services'
+import Story from '../views/story'
 import Tutorial from '../views/Tutorial'
 
 injectTapEventPlugin()
