+++ src/views/logosPosters.js
